removed unused files created new inspection button and connected it to start inspection fragmentjcd
