+++ app/src/main/java/com/example/foundations/AppActivity.java
@@ -90,7 +90,7 @@ public class AppActivity extends AppCompatActivity implements FragmentSwitcher {
                 return true;
             }
         });
-        Fragment dashFragment = new DashFragment();
+        Fragment dashFragment = new DashFragment(this);
         loadFragment(dashFragment);
     }
