+++ app/src/main/res/values/strings.xml
@@ -58,4 +58,6 @@
     <string name="select_inspection">Select Inspection</string>
     <string name="start_new_inspection">Start New Inspection</string>
     <string name="please_select_insp">Please select a profile</string>
+    <string name="new_inspection">NEW INSPECTION</string>
+    <string name="inspection">INSPECTION</string>
 </resources>
