--- strings.xml
@@ -1,63 +1,0 @@
-<resources>
-    <string name="app_name">Foundations</string>
-    <string name="foundations">FOUNDATIONS</string>
-    <string name="home_inspection_made_easy">HOME INSPECTION MADE EASY</string>
-    <string name="get_started">GET STARTED</string>
-    <string name="sign_up">SIGN UP</string>
-    <string name="welcome">WELCOME</string>
-    <string name="license_number">License Number</string>
-    <string name="username">Username</string>
-    <string name="company_name">Company Name</string>
-    <string name="email">Email</string>
-    <string name="phone">Phone</string>
-    <string name="users">USERS</string>
-    <string name="new_user">New User?</string>
-    <string name="continue_as_guest">SELECT USER</string>
-    <string name="empty_field">Please check for empty or Error fields</string>
-    <string name="profile">Profile</string>
-    <string name="inspections">Inspections</string>
-    <string name="checklists">Checklists</string>
-    <string name="pdfs">PDFs</string>
-    <string name="settings">Settings</string>
-    <string name="logout">Logout</string>
-    <string name="user_email">user@foundations.com</string>
-    <string name="open">Open</string>
-    <string name="close">Close</string>
-    <string name="userProfile">currentUserProfile</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
-    <string name="basic_information">Basic Information</string>
-    <string name="buyer_full_name">Buyer Full Name</string>
-    <string name="seller_full_name">Seller Full Name</string>
-    <string name="site_address">Site Address</string>
-    <string name="zip_code">Zip Code</string>
-    <string name="city">City</string>
-    <string name="state">State</string>
-    <string name="buyer_information">Buyer Information</string>
-    <string name="address">Address</string>
-    <string name="seller_information">Seller Information</string>
-    <string name="site_information">Site Information</string>
-    <string name="first_name">First Name</string>
-    <string name="last_name">Last Name</string>
-    <string name="save">SAVE</string>
-    <string name="_98133">98133</string>
-    <string name="wa">WA</string>
-    <string name="seattle">Seattle</string>
-    <string name="michael">Michael</string>
-    <string name="jobs">Jobs</string>
-    <string name="save_and_continue">SAVE AND CONTINUE</string>
-    <string name="introduction">Introduction</string>
-    <string name="exterior">Exterior</string>
-    <string name="interior">Interior</string>
-    <string name="limitations">Limitations</string>
-    <string name="overview">Overview</string>
-    <string name="windows">Windows</string>
-    <string name="doors">Doors</string>
-    <string name="recent_inspections">Recent Inspections</string>
-    <string name="all_inspections">Inspections</string>
-    <string name="select_inspection">Select Inspection</string>
-    <string name="start_new_inspection">Start New Inspection</string>
-    <string name="please_select_insp">Please select a profile</string>
-    <string name="new_inspection">NEW INSPECTION</string>
-    <string name="inspection">INSPECTION</string>
-</resources>
